vroom

+++ info111/arbres-hauteur.cpp
@@ -16,7 +16,7 @@ using namespace std;
 
 int main() {
     vector<vector<string>> t = litTableauCSV("donnees/arbresremarquablesparis.csv", 35);
-    double hauteur_max = 0.0;
+    int hauteur_max = 0;
     int index_max = 0;
     vector<string> C = colonne(t, 8);
     vector<int> tableau_hauteur = conversionInt(C);
